requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.7
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-mathfilters==1.0.0
django-modeltranslation==0.18.4
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
kombu==5.2.4
modeltranslation==0.25
oauthlib==3.2.0
Pillow==9.2.0
postgres==4.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
